requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
psycopg2==2.6
pytz==2014.10
static3==0.5.1
Unipath==1.0
